Clarify unstable PR mergeability state
A PR can still be mergeable while failed checks make it not ready to land.
Show that state as a warning so it is not conflated with a clean merge path.
